1.3.5 修复SetFloat转换位数的错误
* 修复SetFloat转换位数的错误
int -> long

+++ LinePutScript/Line.cs
@@ -631,7 +631,7 @@ namespace LinePutScript
         /// </summary>
         /// <param name="subName">用于定义匹配的名称</param>
         /// <param name="value">储存进sub的double(long)值</param>
-        public void SetFloat(string subName, double value) => FindorAdd(subName).InfoToInt64 = (int)(value * 1000000000);
+        public void SetFloat(string subName, double value) => FindorAdd(subName).InfoToInt64 = (long)(value * 1000000000);
 
         /// <summary>
         /// 获得DateTime属性的sub
